requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
binary==1.0.2
certifi==2025.6.15
frozenlist==1.7.0
golang==0.0.3
greenlet==3.2.3
idna==3.10
importlib_metadata==8.7.0
loguru==0.7.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
propcache==0.3.2
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.1.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.14.0
uuid==1.30
yarl==1.20.1
yoyo-migrations==9.0.0
zipp==3.23.0
